\name{SRRS-package}
\alias{SRRS-package}
\docType{package}
\title{The Stepwise Response Refinement Screener (SRRS)}
\description{A graphical user interface (GUI) implementation of the SRRS method}
\details{
\tabular{ll}{ 
Package: \tab SRRS \cr
Version: \tab 0.1 \cr
Date: \tab 2013-08-05 \cr
License: \tab GPL-2 \cr 
}
The package contains the function \code{SRRS()}, which is the main GUI program for the SRRS method. }
\author{ 
Frederick Kin Hing Phoa

Maintainer: Frederick Kin Hing Phoa <fredphoa@stat.sinica.edu.tw> 
}
\references{ 
1. Phoa, F.K.H. (2013). The Stepwise Response Refinement Screener (SRRS). Accepted by Statistica Sinica.

2. Phoa, F.K.H. (2013). SRRS: An R Package for the Analysis of Factorial and Supersaturated Designs. Submitted
}
\seealso{
\code{\link{SRRS}}
}
\examples{
SRRS()
}
